+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 master
 ------
 
+0.0.12
+------
+
 * Add --type/t flag to "actions" subcommand
 * Add fiat amount value to "actions" subcommand
 * Can give path to config file via option and environment variable